return cache to long

+++ compiler/src/jdk.internal.vm.compiler/src/org/graalvm/compiler/hotspot/meta/Bubo/BuboCache.java
@@ -9,11 +9,11 @@ import java.lang.Thread;
 
  public class BuboCache extends Thread {
         
-        public static int[] Buffer;
+        public static long[] Buffer;
         public static int pointer;
 
         public BuboCache() {
-                Buffer = new int[200_000];
+                Buffer = new long[200_000];
                 pointer = 1;
         }
